streamingProducer/requirements.txt
confluent_kafka
kafka-python
faker
requests
pandas
pyarrow